feat(web): extract appearance settings to a separate tab
Move Theme and Time format settings from the General tab into a new
dedicated Appearance tab, improving settings discoverability.
