+++ apps/web/src/routes/settings.appearance.tsx
@@ -1,0 +1,7 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+import { AppearanceSettingsPanel } from "../components/settings/AppearanceSettingsPanel";
+
+export const Route = createFileRoute("/settings/appearance")({
+  component: AppearanceSettingsPanel,
+});
